+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Add option to age dropped items. Might try to make more specific in the future.
  - Tag spell adds to undo queue.
  - Undo system (and clone+replicate) handle signs, chests, and command blocks. Remove those from the indestructible list.
+ - Don't allow building anything at all from the building list. Affects copy, clone and replicate.
 
 ## 2.8.9.1